key-up and key-down events are now handled
This was a massive rewrite which changed everything. Now instead of just
printing characters out immediately as they're matched, key-up and
key-down events are generated.

The hardware specific code was also factored out into its own hardware
manager class. This massively reduced the code in the key state manager
class which will make it much easier for someone to edit the important
parts.

However, this did break the neat android-esque way of handling keys like
'shift'. So that will have to be fixed.